+++ mrmustard/lab/transformations/mzgate.py
@@ -72,8 +72,12 @@ class MZgate(Unitary):
         internal: bool = False,
     ):
         super().__init__(name="MZgate")
-        self.parameters.add_parameter(make_parameter(phi_a_trainable, phi_a, "phi_a", phi_a_bounds))
-        self.parameters.add_parameter(make_parameter(phi_b_trainable, phi_b, "phi_b", phi_b_bounds))
+        self.parameters.add_parameter(
+            make_parameter(phi_a_trainable, phi_a, "phi_a", phi_a_bounds, dtype=float)
+        )
+        self.parameters.add_parameter(
+            make_parameter(phi_b_trainable, phi_b, "phi_b", phi_b_bounds, dtype=float)
+        )
 
         self._ansatz = PolyExpAnsatz.from_function(
             fn=lambda phi_a, phi_b, internal: Unitary.from_symplectic(
