+++ mrmustard/lab/transformations/s2gate.py
@@ -80,10 +80,14 @@ class S2gate(Unitary):
     ):
         super().__init__(name="S2gate")
         self.parameters.add_parameter(
-            make_parameter(is_trainable=r_trainable, value=r, name="r", bounds=r_bounds),
+            make_parameter(
+                is_trainable=r_trainable, value=r, name="r", bounds=r_bounds, dtype=float
+            ),
         )
         self.parameters.add_parameter(
-            make_parameter(is_trainable=phi_trainable, value=phi, name="phi", bounds=phi_bounds),
+            make_parameter(
+                is_trainable=phi_trainable, value=phi, name="phi", bounds=phi_bounds, dtype=float
+            ),
         )
         self._ansatz = PolyExpAnsatz.from_function(
             fn=triples.twomode_squeezing_gate_Abc,
